New clusterwide operator and CRD
This PR replaces the existing namespaced operator that used
`install.tekton.dev` CRD in favour of a cluster-wide operator that
 relies on a cluster-wide CRD `config.operator.tekton.dev`

The new operator introduces a behavioural change that all OpenShift
pipeline resources will be installed as soon as the operator is
installed/subscribed. This is done by creating an instance of the
`Config` CRD with name - `cluster` and installing pipeline to the
 namespace defined by `spec.targetNamespace` which defaults to
`openshift-pipelines`.

The pipelines can be deleted by deleting the `cluster` CR.

Signed-off-by: sthaha@redhat.com