Use the correct way to get config cluster 🧗‍♂️
The CR config cluster is a cluster-scoped resource, not a namespaced
resource. We need to use the types.NamespacedName without namespace
to really get it.

The check of ignoreRequest is removed, because we need to do further
action, if a deployment of pipelines is deleted.

(cherry picked from commit 588e1d0ff3f16b1454c378aaed8767312caf3598)