Auto creates SA and RBAC for pipelines
Previously, to run pipelines, user was asked to create a SA, configure
RBAC to grant `edit` role. Additional the SA had to be granted
`privileged` SCC. This isn't needed as granting SCC to
pipeline-controller is sufficient to start a privileged container.

This patch addresses the issue above by configuring the `privileged` SCC
and also by auto-configuring SA for all non `openshift` and `kube`
namespaces.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>