Ensure clusterTasks have default values set
Currently operator applies all yaml files in one go and causes the
clusterTasks to be invalid as the webhook may not be active at the time
the task is added and may not get a chance to call defaults.

This is fixed in this patch by doing applying the clusterTasks and
`oc get -o yaml` when the webhook is running, thus ensuring the
defaults are set.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>