Enable usage of tektoncd catalog tasks for s390x and ppc6le
The catalog tasks addition was disabled for s390x and ppc6le
because tasks themselves failed for non Intel architectures.
Now all tasks from the list can be executed on s390x and ppc64le
(in some case container images are multi-arched, for other cases there is
special parameter to specify custom image specific to architecture).
For several tasks version has to be bumped from 0.1 to 0.2, because
custom image parameter is supported only in version 0.2

Signed-off-by: Yulia Gaponenko <yulia.gaponenko1@de.ibm.com>